Implement a Go Spark CLI
For now this focuses on being a drop-in replacement for the current
CLI, where an end user shouldn't notice any difference, except that
the supported 'spark-submit' args now show up in the CLI's --help.

The implementation provides the same arguments that spark-submit uses,
but just converts those into a POST request that's sent to the
dispatcher. This doesn't support things like maven lookup of jar
dependencies, but those things were only applicable to client mode, or
things like YARN, anyway.

This implementation supports both the Binary CLI in 1.8+, while also
providing a simple python wrapper package for backwards compatibility
for 1.7. The stub-universe packaging will automatically fill in the
required binary CLI SHAs for us.