Spark History Server healthcheck update from MESOS_HTTP to COMMAND (#433)
MESSOS_HTTP health check stopped working with Spark 2.3 History Server which binds to a private ip instead of 0.0.0.0 or 127.0.0.1. This resulted in Spark History Server healthcheck failures in Marathon which prevented it from starting. This patch switches from using of MESOS_HTTP to COMMAND health check which greps server logs for errors.