[DCOS-44632 ] Use checksums in `make test-jars` to reduce CI build time (#449)
To reduce CI build times this commit:
- updates the Makefile target to download a pre-built jar and compare
checksums instead of always running `sbt assembly`. This is because the
source code does not change often.
- no longer builds the Lightbend mesos-integration jar as we have
not been running those tests for some time.
- does some README cleanup

This approach does a checksum against all src files and generates
a master checksum. When it comes time to build the jar, we check S3 for
any jars with this checksum and download it otherwise we build the jar,
tag it and upload it so that the next build can benefit.