[DCOS-58441] Allow installing stubs for the infra components (#539)
* Allow installing stubs for the infra components

For MWT20, we'll need to install stubs for the infrastructure components
of the test. This change allows for automating it as part of the script
vs doing it manually.

Also updated the last MWT config file as that will likely be used as
a template for the next test.

* Create section for package repositories.

Also, preserve actual Spark version used for historical purposes.